+++ ChangeLog.md
@@ -2,6 +2,7 @@
 
 ## 2.0.1-dev (changes since 2.0.1 go here)
 
+ * Updated subsample function to be compatible with new version of pandas
  * A unified API for sourcetracking with Gibbs sampling, including
    leave-one-out cross-validation, has been created and is accessible as
    ``sourcetracker.gibbs``.
@@ -9,7 +10,9 @@
    the command line interface, and to the public API as ``sourcetracker.gibbs.plot_heatmap``.
  * The per-sink feature assignments are recorded for every run and written to
    the output directory. They are named as ``X.contingency.txt`` where ``X``
-   is the name of a sink.
+	is the name of a sink.
+
+ * Sample with replacement functionality has been added
 
 ## 2.0.1